Replace ArgumentLoader + JavaTypeMap into a new JavaTypeProvider
A static map was historically used (since Duktape Android) to create and store all existing
JavaType instances once.

Now that we have more dynamic functionalities and support generic arguments (e.g. for Array,
Deferred and FunctionX), some functionalities have been "glued" to the existing implementations
via ArgumentLoader and AdditionalData. This implementation is inconvenient to use and is hardly
readable/maintainable.

This changes gets rid of ArgumentLoader and JavaTypeMap and use a "JavaTypeProvider" instead
which creates JavaType instances on demand (those instances are "cheap").

Also part of this change:
- better error handling
- handle more TODOs