Use dexmaker ProxyBuilder to stub AIDL interface Stub
As AIDL callbacks defined in JS must be mocked to an AIDL interface
Stub which is a real (abstract) class, using a standard Java proxy
does not work as we have to also call the methods implemented by
the stub.
Thanksfully, dexmaker ProxyBuilder allows us to create a proxy whose base class
is a real class (not an interface).